edition 3 fix

+++ OrderService/src/main/java/com/cryptic/dto/OrderItemRequest.java
@@ -1,7 +1,9 @@
 package com.cryptic.dto;
 
+import java.math.BigDecimal;
+
 public record OrderItemRequest(
         String name,
         int quantity,
-        double unitPrice
+        BigDecimal unitPrice
 ) {}
